♻️ refactor : 프로필 API에 수정 부분 통합
